+++ tests/test_create_tile_map.py
@@ -1,5 +1,4 @@
 import json
-from pathlib import Path
 
 import numpy as np
 import pytest
@@ -9,32 +8,31 @@ from osgeo import gdal, osr
 
 from opera_disp_tms import create_tile_map
 from opera_disp_tms import utils as ut
+from opera_disp_tms.constants import SCALE_DICT
 
 
 def test_create_bounds_file(tmp_path, geotiff_info):
-    scale_range = [-1, 1]
-
-    create_tile_map.create_bounds_file(geotiff_info, scale_range, tmp_path)
+    create_tile_map.create_bounds_file(geotiff_info, 'displacement', tmp_path)
 
     with open(f'{tmp_path}/extent.json') as f:
         extent_json = json.load(f)
 
     assert extent_json == {
         'extent': [-113, 33, -112, 34],
-        'scale_range': {'range': [-1, 1], 'units': 'm/yr'},
+        'scale_range': {'range': SCALE_DICT['displacement'], 'units': 'm'},
         'EPSG': 3857,
     }
 
-
-def test_bounds_file_scale_is_rounded(tmp_path, geotiff_info):
-    scale_range = [-0.064441680908203, 0.051021575927734]
-
-    create_tile_map.create_bounds_file(geotiff_info, scale_range, tmp_path)
+    create_tile_map.create_bounds_file(geotiff_info, 'velocity', tmp_path)
 
     with open(f'{tmp_path}/extent.json') as f:
         extent_json = json.load(f)
 
-    assert extent_json['scale_range']['range'] == [-0.064, 0.051]
+    assert extent_json == {
+        'extent': [-113, 33, -112, 34],
+        'scale_range': {'range': SCALE_DICT['velocity'], 'units': 'm/yr'},
+        'EPSG': 3857,
+    }
 
 
 @pytest.fixture
@@ -82,11 +80,11 @@ def test_download_geotiffs(tmp_path):
     for tif in object_keys:
         ut.S3_CLIENT.put_object(Bucket=bucketName, Key=tif)
 
-    output_paths = create_tile_map.download_geotiffs(bucketName, prefix)
+    output_paths = create_tile_map.download_geotiffs(bucketName, prefix, dest_dir=tmp_path)
 
     assert len(output_paths) == 3
     assert output_paths == [
-        Path.cwd() / 'my-file1.tif',
-        Path.cwd() / 'my-file2.tif',
-        Path.cwd() / 'my-file3.tif',
+        tmp_path / 'my-file1.tif',
+        tmp_path / 'my-file2.tif',
+        tmp_path / 'my-file3.tif',
     ]