+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0]
+
+### Changed
+* Map scale range is now globally [-0.05, 0.05] for velocity/secant_velocity and [-0.25, 0.25] for displacement.
+
+### Fixed
+* Values beyond bounds of measurement are now correctly clipped to the min/max value instead of set to NaN.
+* Units are now m/yr for velocity/secant_velocity and m for displacement.
+
 ## [0.7.0]
 Strategy to create the Tile Map Service has been updated to create measurement geotiffs for each OPERA frame, rather
 than creating measurement geotiffs for each 1x1 degree tile.